@@ -4,8 +4,8 @@
   "main": "src/index.tsx",
   "dependencies": {
     "@react-spring/web": "*",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3"
   },
   "scripts": {
@@ -21,8 +21,8 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "typescript": "^4.2.3"
   }
 }
