@@ -6,8 +6,8 @@
     "@react-spring/web": "*",
     "lodash.clamp": "4.0.3",
     "lodash-move": "1.1.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3",
     "react-use-gesture": "^9.1.3"
   },
@@ -25,8 +25,8 @@
   ],
   "devDependencies": {
     "@types/lodash-es": "^4.17.4",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "typescript": "^4.2.3"
   }
 }
