@@ -2,6 +2,7 @@ const ANIMATION_WAIT = 4000
 
 describe('Parallax - vertical', () => {
   const HEIGHT = Cypress.config('viewportHeight')
+  const WIDTH = Cypress.config('viewportWidth')
 
   beforeEach(() => {
     cy.visit('/vertical')
@@ -9,7 +10,11 @@ describe('Parallax - vertical', () => {
 
   it('should translate layers as expected', () => {
     // intial snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    console.log(HEIGHT, WIDTH)
+
+    cy.findByTestId('container').matchImageSnapshot('vertical #1', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // initial layer positions
     cy.findByTestId('default-layer')
@@ -37,7 +42,9 @@ describe('Parallax - vertical', () => {
     cy.wait(ANIMATION_WAIT)
 
     // halfway snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    cy.findByTestId('container').matchImageSnapshot('vertical #2', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // halfway layer positions
     cy.findByTestId('default-layer')
@@ -65,7 +72,9 @@ describe('Parallax - vertical', () => {
     cy.wait(ANIMATION_WAIT)
 
     // final snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    cy.findByTestId('container').matchImageSnapshot('vertical #3', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // final layer positions
     cy.findAllByTestId(/-layer/).each(layer =>
@@ -83,6 +92,7 @@ describe('Parallax - vertical', () => {
 })
 
 describe('Parallax - horizontal', () => {
+  const HEIGHT = Cypress.config('viewportHeight')
   const WIDTH = Cypress.config('viewportWidth')
 
   beforeEach(() => {
@@ -91,7 +101,9 @@ describe('Parallax - horizontal', () => {
 
   it('should translate layers as expected', () => {
     // intial snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    cy.findByTestId('container').matchImageSnapshot('horizontal #1', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // initial layer positions
     cy.findByTestId('default-layer')
@@ -119,7 +131,9 @@ describe('Parallax - horizontal', () => {
     cy.wait(ANIMATION_WAIT)
 
     // halfway snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    cy.findByTestId('container').matchImageSnapshot('horizontal #2', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // halfway layer positions
     cy.findByTestId('default-layer')
@@ -148,7 +162,9 @@ describe('Parallax - horizontal', () => {
     cy.wait(ANIMATION_WAIT)
 
     // final snapshot
-    cy.findByTestId('container').toMatchSnapshot()
+    cy.findByTestId('container').matchImageSnapshot('horizontal #3', {
+      clip: { x: 0, y: 0, width: WIDTH, height: HEIGHT },
+    })
 
     // final layer positions
     cy.findAllByTestId(/-layer/).each(layer =>
