@@ -1,6 +1,6 @@
-const { initPlugin } = require('cypress-plugin-snapshots/plugin')
+const { addMatchImageSnapshotPlugin } = require('cypress-image-snapshot/plugin')
 
 module.exports = (on, config) => {
-  initPlugin(on, config)
+  addMatchImageSnapshotPlugin(on, config)
   return config
 }
