@@ -1,14 +1,15 @@
-import { Text, View, Image, ViewProps, TextProps } from 'react-native'
+import {
+  Text,
+  View,
+  Image,
+  ImageProps,
+  ViewProps,
+  TextProps,
+} from 'react-native'
 import { ComponentClass, ReactNode } from 'react'
 
 export const primitives = {
-  View: View as ComponentClass<
-    // @types/react-native forgot to add "children" to the "View" component??
-    ViewProps & { children?: ReactNode }
-  >,
-  Text: Text as ComponentClass<
-    // @types/react-native forgot to add "children" to the "Text" component??
-    TextProps & { children?: ReactNode }
-  >,
-  Image,
+  View: View as ComponentClass<ViewProps & { children?: ReactNode }>,
+  Text: Text as ComponentClass<TextProps & { children?: ReactNode }>,
+  Image: Image as ComponentClass<ImageProps>,
 }
